+++ src/csb/examination/pages/general_information.html
@@ -21,30 +21,23 @@
     </tr>
   </thead>
 {% endmacro %}
-{% macro data_row(label, value) %}
-  <tr>
-    <td>{{ label }}: {{ paper_corrected::value(value) }}</td>
-    <td></td>
-    <td></td>
-  </tr>
-{% endmacro %}
 {% macro submitter_tables(submitter) %}
   <table class="detail-table">
     {{ columns_head("csb.general.personalia"|trans) }}
     <tbody>
-      {{ data_row("person.fields.initials"|trans, submitter.initials) }}
-      {{ data_row("person.fields.last_name"|trans, submitter.last_name) }}
+      {{ paper_corrected::data_row("person.fields.initials"|trans, submitter.initials) }}
+      {{ paper_corrected::data_row("person.fields.last_name"|trans, submitter.last_name) }}
     </tbody>
   </table>
   <table class="detail-table">
     {{ columns_head("person.correspondence_address"|trans) }}
     <tbody>
-      {{ data_row("person.fields.postal_code"|trans, submitter.postal_code) }}
-      {{ data_row("person.fields.house_number"|trans, submitter.house_number) }}
-      {{ data_row("person.fields.street_name"|trans, submitter.street_name) }}
-      {{ data_row("person.fields.locality"|trans, submitter.locality) }}
-      {{ data_row("person.fields.state_or_province"|trans, submitter.state_or_province) }}
-      {{ data_row("person.fields.country"|trans, submitter.country) }}
+      {{ paper_corrected::data_row("person.fields.postal_code"|trans, submitter.postal_code) }}
+      {{ paper_corrected::data_row("person.fields.house_number"|trans, submitter.house_number) }}
+      {{ paper_corrected::data_row("person.fields.street_name"|trans, submitter.street_name) }}
+      {{ paper_corrected::data_row("person.fields.locality"|trans, submitter.locality) }}
+      {{ paper_corrected::data_row("person.fields.state_or_province"|trans, submitter.state_or_province) }}
+      {{ paper_corrected::data_row("person.fields.country"|trans, submitter.country) }}
     </tbody>
   </table>
 {% endmacro %}
@@ -103,8 +96,8 @@
                aria-label="{{ "action.edit"|trans }}"></a>
           </td>
         </tr>
-        {{ data_row("political_group.list_type"|trans, group_info.list_type) }}
-        {{ data_row("problems.no_previous_election_results"|trans, group_info.previous_results) }}
+        {{ paper_corrected::data_row("political_group.list_type"|trans, group_info.list_type) }}
+        {{ paper_corrected::data_row("problems.no_previous_election_results"|trans, group_info.previous_results) }}
       </tbody>
     </table>
     {% if name_authorisations.is_empty() %}
@@ -114,8 +107,8 @@
         <table class="detail-table">
           {{ columns_head(name_authorisation.heading) }}
           <tbody>
-            {{ data_row("political_group.legal_name"|trans, name_authorisation.legal_name) }}
-            {{ data_row("political_group.authorised_agent"|trans, name_authorisation.authorised_agent) }}
+            {{ paper_corrected::data_row("political_group.legal_name"|trans, name_authorisation.legal_name) }}
+            {{ paper_corrected::data_row("political_group.authorised_agent"|trans, name_authorisation.authorised_agent) }}
           </tbody>
         </table>
       {% endfor %}
