+++ src/csb/examination/structs/mod.rs
@@ -1,6 +1,8 @@
 mod csb_candidate;
 mod csb_candidate_list;
 mod paper_corrected;
+mod all_omissions;
+mod all_csb_corrections;
 
 pub use csb_candidate::CsbCandidate;
 pub use csb_candidate_list::CsbCandidateList;
@@ -9,3 +11,5 @@ pub use paper_corrected::{
     PaperCorrectedPoliticalGroupInfo, PaperCorrectedSubmitter, paper_corrected_list_submitter,
     paper_corrected_name_authorisations, paper_corrected_substitute_submitters,
 };
+pub use all_omissions::AllOmissions;
+pub use all_csb_corrections::AllCsbCorrections;
