+++ src/structs/common/date.rs
@@ -8,7 +8,7 @@ use crate::{ElectionConfig, constants::DEFAULT_DATE_FORMAT, form::ValidationErro
 static DATE_FORMAT_REGEX: LazyLock<Regex> =
     LazyLock::new(|| Regex::new(r"^\d{2}-\d{2}-\d{4}$").unwrap());
 
-#[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize)]
+#[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize, Hash)]
 #[serde(transparent)]
 pub struct DateOfBirth(NaiveDate);
